requirements.txt
appdirs==1.4.4
asgiref==3.8.1
atomicwrites==1.4.0
backoff==2.2.1
backports.weakref==1.0.post1
bcrypt==4.1.3
build==1.2.1
chroma-hnswlib==0.7.3
chromadb==0.5.3
clyent==1.2.2
coloredlogs==15.0.1
conda-repo-cli==1.0.75
conda-verify==3.4.2
dataclasses-json==0.6.7
Deprecated==1.2.14
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.111.1
fastapi-cli==0.0.4
flatbuffers==24.3.25
fonttools==4.25.0
glob2==0.7
google-auth==2.32.0
googleapis-common-protos==1.63.2
grpcio==1.65.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.0
humanfriendly==10.0
importlib_resources==6.4.0
inflection==0.5.1
jsonpatch==1.33
jsonpointer==2.1
kubernetes==30.1.0
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-text-splitters==0.2.2
langsmith==0.1.93
marshmallow==3.21.3
mkl-service==2.4.0
mmh3==4.1.0
monotonic==1.6
munkres==1.1.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.37.0
opencv-python==4.10.0.84
openpyxl==3.0.10
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.6
packaging==24.1
patsy==0.5.3
pdf2image==1.17.0
ply==3.11
posthog==3.5.0
protobuf==3.20.3
pyasn1-modules==0.2.8
PyDispatcher==2.0.5
pyls-spyder==0.4.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyPika==0.48.9
pyproject_hooks==1.1.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
pytesseract==0.3.10
python-docx==1.1.2
python-lsp-jsonrpc==1.0.0
python-multipart==0.0.9
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.3
sentence-transformers==3.0.1
shellingham==1.5.4
starlette==0.37.2
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.1+cpu
torchaudio==2.3.1+cpu
torchvision==0.18.1+cpu
transformers==4.42.4
triton==2.3.1
typer==0.12.3
typing-inspect==0.9.0
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
webencodings==0.5.1
websockets==12.0
